+++ GUI/DataManager/CharacterListEntry/CharacterListEntryMixin.lua
@@ -37,81 +37,81 @@ end
 
 function KrowiAF_CharacterListEntryMixin:ToggleHeaderTooltip()
     if self.HeaderTooltip:GetChecked() then
-        SavedData.Characters[self.Guid].ExcludeFromHeaderTooltip = nil;
-        SavedData.Characters[self.Guid].Ignore = nil;
+        KrowiAF_SavedData.Characters[self.Guid].ExcludeFromHeaderTooltip = nil;
+        KrowiAF_SavedData.Characters[self.Guid].Ignore = nil;
 
         self.IgnoreCharacter:SetChecked(false);
     else
-        SavedData.Characters[self.Guid].ExcludeFromHeaderTooltip = true;
+        KrowiAF_SavedData.Characters[self.Guid].ExcludeFromHeaderTooltip = true;
     end
-    self.Character.ExcludeFromHeaderTooltip = SavedData.Characters[self.Guid].ExcludeFromHeaderTooltip;
-    self.Character.IgnoreCharacter = SavedData.Characters[self.Guid].Ignore;
+    self.Character.ExcludeFromHeaderTooltip = KrowiAF_SavedData.Characters[self.Guid].ExcludeFromHeaderTooltip;
+    self.Character.IgnoreCharacter = KrowiAF_SavedData.Characters[self.Guid].Ignore;
 end
 
 function KrowiAF_CharacterListEntryMixin:ToggleEarnedByAchievementTooltip()
     if self.EarnedByAchievementTooltip:GetChecked() then
-        SavedData.Characters[self.Guid].ExcludeFromEarnedByAchievementTooltip = nil;
-        SavedData.Characters[self.Guid].Ignore = nil;
+        KrowiAF_SavedData.Characters[self.Guid].ExcludeFromEarnedByAchievementTooltip = nil;
+        KrowiAF_SavedData.Characters[self.Guid].Ignore = nil;
 
         self.IgnoreCharacter:SetChecked(false);
     else
-        SavedData.Characters[self.Guid].ExcludeFromEarnedByAchievementTooltip = true;
+        KrowiAF_SavedData.Characters[self.Guid].ExcludeFromEarnedByAchievementTooltip = true;
     end
-    self.Character.ExcludeFromEarnedByAchievementTooltip = SavedData.Characters[self.Guid].ExcludeFromEarnedByAchievementTooltip;
-    self.Character.IgnoreCharacter = SavedData.Characters[self.Guid].Ignore;
+    self.Character.ExcludeFromEarnedByAchievementTooltip = KrowiAF_SavedData.Characters[self.Guid].ExcludeFromEarnedByAchievementTooltip;
+    self.Character.IgnoreCharacter = KrowiAF_SavedData.Characters[self.Guid].Ignore;
 end
 
 function KrowiAF_CharacterListEntryMixin:ToggleMostProgressAchievementTooltip()
     if self.MostProgressAchievementTooltip:GetChecked() then
-        SavedData.Characters[self.Guid].ExcludeFromMostProgressAchievementTooltip = nil;
-        SavedData.Characters[self.Guid].Ignore = nil;
+        KrowiAF_SavedData.Characters[self.Guid].ExcludeFromMostProgressAchievementTooltip = nil;
+        KrowiAF_SavedData.Characters[self.Guid].Ignore = nil;
 
         self.IgnoreCharacter:SetChecked(false);
     else
-        SavedData.Characters[self.Guid].ExcludeFromMostProgressAchievementTooltip = true;
+        KrowiAF_SavedData.Characters[self.Guid].ExcludeFromMostProgressAchievementTooltip = true;
     end
-    self.Character.ExcludeFromMostProgressAchievementTooltip = SavedData.Characters[self.Guid].ExcludeFromMostProgressAchievementTooltip;
-    self.Character.IgnoreCharacter = SavedData.Characters[self.Guid].Ignore;
+    self.Character.ExcludeFromMostProgressAchievementTooltip = KrowiAF_SavedData.Characters[self.Guid].ExcludeFromMostProgressAchievementTooltip;
+    self.Character.IgnoreCharacter = KrowiAF_SavedData.Characters[self.Guid].Ignore;
 end
 
 function KrowiAF_CharacterListEntryMixin:ToggleIgnoreCharacter()
     if self.IgnoreCharacter:GetChecked() then
-        SavedData.Characters[self.Guid].Ignore = true;
-        SavedData.Characters[self.Guid].ExcludeFromHeaderTooltip = true;
-        SavedData.Characters[self.Guid].ExcludeFromEarnedByAchievementTooltip = true;
-        SavedData.Characters[self.Guid].ExcludeFromMostProgressAchievementTooltip = true;
-        SavedData.Characters[self.Guid].CompletedAchievements = {};
-        SavedData.Characters[self.Guid].NotCompletedAchievements = {};
-        SavedData.Characters[self.Guid].LastCompleted = nil;
-        SavedData.Characters[self.Guid].Points = 0;
+        KrowiAF_SavedData.Characters[self.Guid].Ignore = true;
+        KrowiAF_SavedData.Characters[self.Guid].ExcludeFromHeaderTooltip = true;
+        KrowiAF_SavedData.Characters[self.Guid].ExcludeFromEarnedByAchievementTooltip = true;
+        KrowiAF_SavedData.Characters[self.Guid].ExcludeFromMostProgressAchievementTooltip = true;
+        KrowiAF_SavedData.Characters[self.Guid].CompletedAchievements = {};
+        KrowiAF_SavedData.Characters[self.Guid].NotCompletedAchievements = {};
+        KrowiAF_SavedData.Characters[self.Guid].LastCompleted = nil;
+        KrowiAF_SavedData.Characters[self.Guid].Points = 0;
 
         self.Points:SetText("0");
         self.HeaderTooltip:SetChecked(false);
         self.EarnedByAchievementTooltip:SetChecked(false);
         self.MostProgressAchievementTooltip:SetChecked(false);
     else
-        SavedData.Characters[self.Guid].Ignore = nil;
-        SavedData.Characters[self.Guid].ExcludeFromHeaderTooltip = nil;
-        SavedData.Characters[self.Guid].ExcludeFromEarnedByAchievementTooltip = nil;
-        SavedData.Characters[self.Guid].ExcludeFromMostProgressAchievementTooltip = nil;
+        KrowiAF_SavedData.Characters[self.Guid].Ignore = nil;
+        KrowiAF_SavedData.Characters[self.Guid].ExcludeFromHeaderTooltip = nil;
+        KrowiAF_SavedData.Characters[self.Guid].ExcludeFromEarnedByAchievementTooltip = nil;
+        KrowiAF_SavedData.Characters[self.Guid].ExcludeFromMostProgressAchievementTooltip = nil;
         if self.Guid == UnitGUID("player") then
             addon.ResetCache();
             addon.BuildCache();
-            self.Points:SetText(SavedData.Characters[self.Guid].Points or "99999");
+            self.Points:SetText(KrowiAF_SavedData.Characters[self.Guid].Points or "99999");
         end
 
         self.HeaderTooltip:SetChecked(true);
         self.EarnedByAchievementTooltip:SetChecked(true);
         self.MostProgressAchievementTooltip:SetChecked(true);
     end
-    self.Character.ExcludeFromHeaderTooltip = SavedData.Characters[self.Guid].ExcludeFromHeaderTooltip;
-    self.Character.ExcludeFromEarnedByAchievementTooltip = SavedData.Characters[self.Guid].ExcludeFromEarnedByAchievementTooltip;
-    self.Character.ExcludeFromMostProgressAchievementTooltip = SavedData.Characters[self.Guid].ExcludeFromMostProgressAchievementTooltip;
-    self.Character.IgnoreCharacter = SavedData.Characters[self.Guid].Ignore;
+    self.Character.ExcludeFromHeaderTooltip = KrowiAF_SavedData.Characters[self.Guid].ExcludeFromHeaderTooltip;
+    self.Character.ExcludeFromEarnedByAchievementTooltip = KrowiAF_SavedData.Characters[self.Guid].ExcludeFromEarnedByAchievementTooltip;
+    self.Character.ExcludeFromMostProgressAchievementTooltip = KrowiAF_SavedData.Characters[self.Guid].ExcludeFromMostProgressAchievementTooltip;
+    self.Character.IgnoreCharacter = KrowiAF_SavedData.Characters[self.Guid].Ignore;
 end
 
 local function DeleteCharacterCallback(self)
-    SavedData.Characters[self.Guid] = nil;
+    KrowiAF_SavedData.Characters[self.Guid] = nil;
     addon.GUI.DataManagerFrame.CharacterList:Refresh();
 end
 
@@ -121,7 +121,7 @@ function KrowiAF_CharacterListEntryMixin:DeleteCharacterFunction()
             {
                 text = addon.L["Are you sure you want to delete character?"]:ReplaceVars
                 {
-                    character = SavedData.Characters[self.Guid].Name:SetColorYellow(),
+                    character = KrowiAF_SavedData.Characters[self.Guid].Name:SetColorYellow(),
                     ignore = addon.L["Ignore"]:SetColorYellow()
                 },
                 callback = function()